move shared toggle component into sharedComponents directory
